+++ NEWS.md
@@ -24,8 +24,8 @@ _This version is in development_
 
 * `univariate_table`
 
-  - Default access to 25th and 75th percentiles for string templates with "q1" and "q3", respectively
-  - Use "median (q1, q3)" as the default string template for numeric variables
+  - Default access to 25th and 75th percentiles for string templates with `"q1"` and `"q3"`, respectively
+  - Use `"median (q1, q3)"` as the default string template for numeric variables
 
 ## New functions